requirements.txt
fpdf==1.7.2
parse==1.12.0
py-dateutil==2.2
six==1.12.0